chore: move in single function
